#2071 Removes WireMockListener extension methods
