+++ kotest-extensions/kotest-extensions-wiremock/src/jvmTest/kotlin/io/kotest/extensions/wiremock/WiremockListenerPerTestTest.kt
@@ -16,7 +16,7 @@ import java.net.URL
 class WiremockListenerPerTestTest : FunSpec({
    val wireMockServer = WireMockServer(9000)
 
-   listener(wireMockServer.listenerPerTest())
+   listener(WireMockListener.perTest(wireMockServer))
    listener(object : TestListener {
       override suspend fun afterTest(testCase: TestCase, result: TestResult) {
          shouldThrow<ConnectException> {
